merge deepmodeling/reconstruction branch update to abacusmodeling/develop

+++ tests/module_base/blas_connector.cpp
@@ -1,17 +1,17 @@
-#include "gtest/gtest.h"
 #include "module_base/blas_connector.h"
+#include "gtest/gtest.h"
 
-#include <array>
-#include <cstdlib>
 #include <algorithm>
+#include <array>
 #include <complex>
+#include <cstdlib>
 TEST(blas_connector, sscal_)
 {
 	typedef float T;
 	const int size = 8;
 	const T scale = 2;
 	const int incx = 1;
-	std::array<T,size> result, answer;
+	std::array<T, size> result, answer;
 	std::generate(result.begin(), result.end(), std::rand);
 	for (int i = 0; i < size; i++)
 		answer[i] = result[i] * scale;
@@ -26,7 +26,7 @@ TEST(blas_connector, dscal_)
 	const int size = 8;
 	const T scale = 2;
 	const int incx = 1;
-	std::array<T,size> result, answer;
+	std::array<T, size> result, answer;
 	std::generate(result.begin(), result.end(), std::rand);
 	for (int i = 0; i < size; i++)
 		answer[i] = result[i] * scale;
@@ -39,15 +39,17 @@ TEST(blas_connector, cscal_)
 {
 	typedef std::complex<float> T;
 	const int size = 8;
-	const T scale = {2,3};
+	const T scale = {2, 3};
 	const int incx = 1;
-	std::array<T,size> result, answer;
+	std::array<T, size> result, answer;
 	std::generate(result.begin(), result.end(), []()
-				  { return T{std::rand(), std::rand()}; });
+				  { return T{static_cast<float>(std::rand()),
+							 static_cast<float>(std::rand())}; });
 	for (int i = 0; i < size; i++)
 		answer[i] = result[i] * scale;
 	cscal_(&size, &scale, result.data(), &incx);
-	for (int i = 0; i < size; i++){
+	for (int i = 0; i < size; i++)
+	{
 		EXPECT_FLOAT_EQ(answer[i].real(), result[i].real());
 		EXPECT_FLOAT_EQ(answer[i].imag(), result[i].imag());
 	}
@@ -61,7 +63,8 @@ TEST(blas_connector, zscal_)
 	const int incx = 1;
 	std::array<T, size> result, answer;
 	std::generate(result.begin(), result.end(), []()
-				  { return T{std::rand(), std::rand()}; });
+				  { return T{static_cast<double>(std::rand()),
+							 static_cast<double>(std::rand())}; });
 	for (int i = 0; i < size; i++)
 		answer[i] = result[i] * scale;
 	zscal_(&size, &scale, result.data(), &incx);
